Fixing issue with 'Value too large'

+++ Apps.Monday/Webhooks/ItemWebhookList.cs
@@ -73,7 +73,7 @@ public class ItemWebhookList(InvocationContext invocationContext) : AppInvocable
 
     private async Task<ItemResponse> GetItemAsync(string itemId)
     {
-        var variables = new { ids = int.Parse(itemId) };
+        var variables = new { ids = long.Parse(itemId) };
         var request = new ApiRequest(GraphQlQueries.GetItemById, variables, Creds);
 
         var response = await Client.ExecuteWithErrorHandling<DataWrapperDto<SearchItemsResponse>>(request);
